core - remove py2.7 compat vendored ipaddress module (#5479)
