+++ c7n/utils.py
@@ -18,6 +18,7 @@ import csv
 from datetime import datetime, timedelta
 import json
 import itertools
+import ipaddress
 import logging
 import os
 import random
@@ -30,7 +31,7 @@ import six
 from six.moves.urllib import parse as urlparse
 from six.moves.urllib.request import getproxies
 
-from c7n import ipaddress, config
+from c7n import config
 from c7n.exceptions import ClientError, PolicyValidationError
 
 # Try to play nice in a serverless environment, where we don't require yaml
@@ -437,6 +438,23 @@ class IPv4Network(ipaddress.IPv4Network):
         if isinstance(other, ipaddress._BaseNetwork):
             return self.supernet_of(other)
         return super(IPv4Network, self).__contains__(other)
+
+    if (sys.version_info.major == 3 and sys.version_info.minor <= 6):  # pragma: no cover
+        @staticmethod
+        def _is_subnet_of(a, b):
+            try:
+                # Always false if one is v4 and the other is v6.
+                if a._version != b._version:
+                    raise TypeError(f"{a} and {b} are not of the same version")
+                return (b.network_address <= a.network_address and
+                        b.broadcast_address >= a.broadcast_address)
+            except AttributeError:
+                raise TypeError(f"Unable to test subnet containment "
+                                f"between {a} and {b}")
+
+        def supernet_of(self, other):
+            """Return True if this network is a supernet of other."""
+            return self._is_subnet_of(other, self)
 
 
 def reformat_schema(model):
