+++ pyproject.toml
@@ -41,7 +41,6 @@ vcrpy = "^4.0.2"
 aws-xray-sdk = "^2.4.3"
 flake8 = "^3.7.9"
 tox = "^3.14.5"
-pytest-sugar = "^0.9.2"
 jsonpatch = "^1.25"
 psutil = "^5.7.0"
 # wheel = "==0.30.0"
